move txrx to mogwai-chan, patch list model
