+++ Cargo.toml
@@ -2,6 +2,8 @@
 
 members = [
   "cookbook",
+  "crates/gremlin",
+  "crates/mogwai-chan",
   "crates/mogwai-html-macro",
   "crates/mogwai-hydrator",
   "examples/focus-follower",
