+++ crates/mogwai-chan/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "mogwai-chan"
+version = "0.1.0"
+authors = ["Schell Carl Scivally <efsubenovex@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
