+++ mogwai/Cargo.toml
@@ -12,6 +12,7 @@ keywords = ["ui", "dom", "app", "reactive", "frontend"]
 categories = ["gui", "wasm", "web-programming"]
 
 [dependencies]
+mogwai-chan = { path = "../crates/mogwai-chan", version = "^0.1" }
 console_log = "^0.1"
 log = "^0.4"
 js-sys = "0.3.44"
